# autoevolve-ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Company Cockpit",
    "version": "0.1.0"
  },
  "paths": {
...
    "/company/ui/onboarding": {
      "get": {
        "tags": [
          "onboarding"
        ],
        "summary": "Onboarding View",
        "operationId": "onboarding_view_company_ui_onboarding_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingView"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
...
    "/company/setup/onboarding/company": {
      "post": {
        "tags": [
          "onboarding"
        ],
        "summary": "Start Company",
        "operationId": "start_company_company_setup_onboarding_company_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "x-founder-action-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Founder-Action-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyStart"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/company/setup/onboarding/research-company": {
      "post": {
        "tags": [
          "onboarding"
        ],
        "summary": "Research Company",
        "operationId": "research_company_company_setup_onboarding_research_company_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "x-founder-action-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Founder-Action-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/company/setup/onboarding/company/confirm": {
      "post": {
        "tags": [
          "onboarding"
        ],
        "summary": "Confirm Company",
        "operationId": "confirm_company_company_setup_onboarding_company_confirm_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "x-founder-action-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Founder-Action-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyContext"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/company/setup/onboarding/strategy": {
      "post": {
        "tags": [
          "onboarding"
        ],
        "summary": "Confirm Strategy",
        "operationId": "confirm_strategy_company_setup_onboarding_strategy_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "x-founder-action-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Founder-Action-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/company/setup/onboarding/search": {
      "post": {
        "tags": [
          "onboarding"
        ],
        "summary": "First Search",
        "operationId": "first_search_company_setup_onboarding_search_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "x-founder-action-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Founder-Action-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/company/setup/onboarding/calibration": {
      "post": {
        "tags": [
          "onboarding"
        ],
        "summary": "Record Calibration",
        "operationId": "record_calibration_company_setup_onboarding_calibration_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "x-founder-action-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Founder-Action-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalibrationInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/company/setup/onboarding/refinement": {
      "post": {
        "tags": [
          "onboarding"
        ],
        "summary": "Decide Refinement",
        "operationId": "decide_refinement_company_setup_onboarding_refinement_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "x-founder-action-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Founder-Action-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefinementDecision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CalibrationInput": {
        "properties": {
          "feedback": {
            "items": {
              "$ref": "#/components/schemas/FeedbackEntry"
            },
            "type": "array",
            "maxItems": 10,
            "minItems": 1,
            "title": "Feedback"
          }
        },
        "type": "object",
        "required": [
          "feedback"
        ],
        "title": "CalibrationInput"
      },
...
      "CompanyStart": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Name"
          },
          "website": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Website"
          },
          "objective": {
            "type": "string",
            "maxLength": 300,
            "minLength": 3,
            "title": "Objective"
          },
          "market": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Market"
          }
        },
        "type": "object",
        "required": [
          "name",
          "website",
          "objective",
          "market"
        ],
        "title": "CompanyStart"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OnboardingView": {
        "properties": {
          "program": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProgramState"
              },
              {
                "type": "null"
              }
            ]
          },
          "next_step": {
            "type": "string",
            "title": "Next Step"
          },
          "sample": {
            "items": {
              "$ref": "#/components/schemas/CalibrationCandidate"
            },
            "type": "array",
            "title": "Sample"
          },
          "confidence_basis": {
            "type": "string",
            "title": "Confidence Basis",
            "default": "Existing lead_score only; not a verified ICP fit score."
          }
        },
        "type": "object",
        "required": [
          "program",
          "next_step",
          "sample"
        ],
        "title": "OnboardingView"
      },
...
      "RefinementDecision": {
        "properties": {
          "approved": {
            "type": "boolean",
            "title": "Approved"
          }
        },
        "type": "object",
        "required": [
          "approved"
        ],
        "title": "RefinementDecision"
      },
...
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}